% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.hdsvm-methods.R
\name{coef.cv.hdsvm}
\alias{coef.cv.hdsvm}
\title{Extract Coefficients from a \code{cv.hdsvm} Object}
\usage{
\method{coef}{cv.hdsvm}(object, s = c("lambda.1se", "lambda.min"), ...)
}
\arguments{
\item{object}{A fitted \code{\link{cv.hdsvm}} object.}

\item{s}{Value(s) of the penalty parameter \code{lambda} at which
coefficients are desired. The default is \code{s = "lambda.1se"}, the
largest value of \code{lambda} such that the cross-validation error is
within one standard error of the minimum. Alternatively,
\code{s = "lambda.min"} gives the value minimizing the cross-validation
error. If \code{s} is numeric, these are taken as the actual values of
\code{lambda} to use.}

\item{...}{Not used.}
}
\value{
A matrix of coefficients at the specified \code{lambda} values.
}
\description{
Retrieves coefficients from a cross-validated \code{\link{hdsvm}} model,
using the stored full-data fit and the value of \code{lambda} selected by
cross-validation.
}
\examples{
set.seed(315)
n <- 100
p <- 400
x1 <- matrix(rnorm(n / 2 * p, -0.25, 0.1), n / 2)
x2 <- matrix(rnorm(n / 2 * p, 0.25, 0.1), n / 2)
x <- rbind(x1, x2)
beta <- 0.1 * rnorm(p)
prob <- plogis(c(x \%*\% beta))
y <- 2 * rbinom(n, 1, prob) - 1
cv.fit <- cv.hdsvm(x, y, lam2 = 0.01)
coef(cv.fit, s = c(0.02, 0.03))
}
\seealso{
\code{\link{cv.hdsvm}}, \code{\link{predict.cv.hdsvm}}
}
