% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huber-methods.R
\name{coef.hdhuber}
\alias{coef.hdhuber}
\title{Extract Model Coefficients from a \code{hdhuber} Object}
\usage{
\method{coef}{hdhuber}(object, s = NULL, type = c("coefficients", "nonzero"), ...)
}
\arguments{
\item{object}{Fitted \code{\link{hdhuber}} object.}

\item{s}{Values of the penalty parameter \code{lambda} at which
coefficients are requested. Default is the entire sequence used during
the model fit.}

\item{type}{Type \code{"coefficients"} (the default) returns the
coefficients at the requested values of \code{s}; type \code{"nonzero"}
returns a list of the indices of the nonzero coefficients for each value
of \code{s}.}

\item{...}{Not used.}
}
\value{
A matrix of coefficients (intercept in the first row, one column
  per value of \code{s}), or a list of indices when
  \code{type = "nonzero"}.
}
\description{
Retrieves the coefficients at specified values of \code{lambda} from a
fitted \code{\link{hdhuber}} object. If \code{s}, the vector of
\code{lambda} values, contains values not used in the model fitting,
linear interpolation between the closest fitted \code{lambda} values is
used.
}
\examples{
set.seed(315)
n <- 100
p <- 400
x <- matrix(data = rnorm(n * p, mean = 0, sd = 1), nrow = n, ncol = p)
beta_star <- c(c(2, 1.5, 0.8, 1, 1.75, 0.75, 0.3), rep(0, (p - 7)))
eps <- rnorm(n, mean = 0, sd = 1)
y <- x \%*\% beta_star + eps
fit <- hdhuber(x, y, delta = 0.5, lam2 = 0)
coefs <- coef(fit, s = fit$lambda[3:5])
}
\seealso{
\code{\link{hdhuber}}, \code{\link{predict.hdhuber}}
}
