% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdsvm-methods.R
\name{coef.hdsvm}
\alias{coef.hdsvm}
\title{Extract Model Coefficients from a \code{hdsvm} Object}
\usage{
\method{coef}{hdsvm}(object, s = NULL, type = c("coefficients", "nonzero"), ...)
}
\arguments{
\item{object}{Fitted \code{\link{hdsvm}} object.}

\item{s}{Values of the penalty parameter \code{lambda} at which
coefficients are requested. Default is the entire sequence used during
the model fit.}

\item{type}{Type \code{"coefficients"} (the default) returns the
coefficients at the requested values of \code{s}; type \code{"nonzero"}
returns a list of the indices of the nonzero coefficients for each value
of \code{s}.}

\item{...}{Not used.}
}
\value{
A matrix of coefficients (intercept in the first row, one column
  per value of \code{s}), or a list of indices when
  \code{type = "nonzero"}.
}
\description{
Retrieves the coefficients at specified values of \code{lambda} from a
fitted \code{\link{hdsvm}} object. If \code{s}, the vector of
\code{lambda} values, contains values not used in the model fitting,
linear interpolation between the closest fitted \code{lambda} values is
used.
}
\examples{
set.seed(315)
n <- 100
p <- 400
x1 <- matrix(rnorm(n / 2 * p, -0.25, 0.1), n / 2)
x2 <- matrix(rnorm(n / 2 * p, 0.25, 0.1), n / 2)
x <- rbind(x1, x2)
beta <- 0.1 * rnorm(p)
prob <- plogis(c(x \%*\% beta))
y <- 2 * rbinom(n, 1, prob) - 1
fit <- hdsvm(x, y, lam2 = 0.01)
coefs <- coef(fit, s = fit$lambda[3:5])
}
\seealso{
\code{\link{hdsvm}}, \code{\link{predict.hdsvm}}
}
