% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{plot.cv}
\alias{plot.cv}
\alias{plot.cv.hdhuber}
\alias{plot.cv.hdsvm}
\alias{plot.cv.hdqr}
\alias{plot.cv.nc.hdsvm}
\alias{plot.cv.nc.hdqr}
\title{Plot the Cross-validation Curve}
\usage{
\method{plot}{cv.hdhuber}(x, sign.lambda = 1, ...)

\method{plot}{cv.hdsvm}(x, sign.lambda = 1, ...)

\method{plot}{cv.hdqr}(x, sign.lambda = 1, ...)

\method{plot}{cv.nc.hdsvm}(x, sign.lambda = 1, ...)

\method{plot}{cv.nc.hdqr}(x, sign.lambda = 1, ...)
}
\arguments{
\item{x}{A fitted \code{cv.hdhuber}, \code{cv.hdsvm}, \code{cv.hdqr},
\code{cv.hdrr}, \code{cv.nc.hdsvm}, or \code{cv.nc.hdqr} object.}

\item{sign.lambda}{Either plot against \code{log(lambda)} (default) or
its negative if \code{sign.lambda = -1}.}

\item{...}{Other graphical parameters passed to \code{plot}.}
}
\value{
No return value; called for its side effect of producing a plot.
}
\description{
Plots the cross-validation curve, with upper and lower standard-error
bars, as a function of \code{log(lambda)} for a cross-validated fit.
Vertical dotted lines mark \code{lambda.min} and \code{lambda.1se}, and
the number of nonzero coefficients is shown along the top axis.
}
\examples{
set.seed(315)
n <- 100
p <- 400
x <- matrix(data = rnorm(n * p, mean = 0, sd = 1), nrow = n, ncol = p)
beta_star <- c(c(2, 1.5, 0.8, 1, 1.75, 0.75, 0.3), rep(0, (p - 7)))
eps <- rnorm(n, mean = 0, sd = 1)
y <- x \%*\% beta_star + eps
cv.fit <- cv.hdqr(x = x, y = y, tau = 0.5)
plot(cv.fit)
}
\seealso{
\code{\link{cv.hdhuber}}, \code{\link{cv.hdsvm}},
  \code{\link{cv.hdqr}}, \code{\link{cv.hdrr}}
}
