% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvqr-methods.R
\name{predict.cv.hdqr}
\alias{predict.cv.hdqr}
\title{Make Predictions from a \code{cv.hdqr} Object}
\usage{
\method{predict}{cv.hdqr}(object, newx, s = c("lambda.1se", "lambda.min"), ...)
}
\arguments{
\item{object}{A fitted \code{\link{cv.hdqr}} or \code{\link{cv.hdrr}}
object.}

\item{newx}{Matrix of new predictor values at which predictions are to be
made. This is a required argument.}

\item{s}{Value(s) of the penalty parameter \code{lambda} at which
predictions are desired. The default is \code{s = "lambda.1se"}, the
largest value of \code{lambda} such that the cross-validation error is
within one standard error of the minimum. Alternatively,
\code{s = "lambda.min"} gives the value minimizing the cross-validation
error. If \code{s} is numeric, these are taken as the actual values of
\code{lambda} to use.}

\item{...}{Not used.}
}
\value{
A matrix of predicted values, one row per row of \code{newx} and
  one column per value of \code{s}.
}
\description{
Generates predictions from a cross-validated \code{\link{hdqr}} (or
\code{\link{hdrr}}) model, using the stored full-data fit and the value
of \code{lambda} selected by cross-validation.
}
\examples{
set.seed(315)
n <- 100
p <- 400
x <- matrix(data = rnorm(n * p, mean = 0, sd = 1), nrow = n, ncol = p)
beta_star <- c(c(2, 1.5, 0.8, 1, 1.75, 0.75, 0.3), rep(0, (p - 7)))
eps <- rnorm(n, mean = 0, sd = 1)
y <- x \%*\% beta_star + eps
cv.fit <- cv.hdqr(x = x, y = y, tau = 0.5, lam2 = 0.01)
predict(cv.fit, newx = x[50:60, ], s = "lambda.min")
}
\seealso{
\code{\link{cv.hdqr}}, \code{\link{coef.cv.hdqr}}
}
