% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr-methods.R
\name{predict.hdqr}
\alias{predict.hdqr}
\title{Make Predictions from a \code{hdqr} Object}
\usage{
\method{predict}{hdqr}(object, newx, s = NULL, ...)
}
\arguments{
\item{object}{Fitted \code{\link{hdqr}} or \code{\link{hdrr}} object.}

\item{newx}{Matrix of new predictor values at which predictions are to be
made. This is a required argument.}

\item{s}{Values of the penalty parameter \code{lambda} at which
predictions are requested. Default is the entire sequence used during
the model fit.}

\item{...}{Not used.}
}
\value{
A matrix of predicted values, one row per row of \code{newx} and
  one column per value of \code{s}.
}
\description{
Produces fitted values for new predictor data using a fitted
\code{\link{hdqr}} (or \code{\link{hdrr}}) object at specified
\code{lambda} values.
}
\examples{
set.seed(315)
n <- 100
p <- 400
x <- matrix(data = rnorm(n * p, mean = 0, sd = 1), nrow = n, ncol = p)
beta_star <- c(c(2, 1.5, 0.8, 1, 1.75, 0.75, 0.3), rep(0, (p - 7)))
eps <- rnorm(n, mean = 0, sd = 1)
y <- x \%*\% beta_star + eps
fit <- hdqr(x = x, y = y, tau = 0.5, lam2 = 0.01)
preds <- predict(fit, newx = tail(x), s = fit$lambda[3:5])
}
\seealso{
\code{\link{hdqr}}, \code{\link{coef.hdqr}}
}
