% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.hdqr-methods.R
\name{predict.nc.hdqr}
\alias{predict.nc.hdqr}
\title{Make Predictions from a \code{nc.hdqr} Object}
\usage{
\method{predict}{nc.hdqr}(object, newx, s = NULL, ...)
}
\arguments{
\item{object}{Fitted \code{\link{nc.hdqr}} object.}

\item{newx}{Matrix of new predictor values at which predictions are to be
made. This is a required argument.}

\item{s}{Values of the penalty parameter \code{lambda} at which
predictions are requested. Default is the entire sequence used during
the model fit.}

\item{...}{Not used.}
}
\value{
A matrix of predicted values, one row per row of \code{newx} and
  one column per value of \code{s}.
}
\description{
Produces fitted values for new predictor data using a fitted
\code{\link{nc.hdqr}} object at specified \code{lambda} values.
}
\examples{
set.seed(315)
n <- 100
p <- 400
x <- matrix(data = rnorm(n * p, mean = 0, sd = 1), nrow = n, ncol = p)
beta_star <- c(c(2, 1.5, 0.8, 1, 1.75, 0.75, 0.3), rep(0, (p - 7)))
eps <- rnorm(n, mean = 0, sd = 1)
y <- x \%*\% beta_star + eps
lambda <- 10^(seq(0, -2, length.out = 20))
nc.fit <- nc.hdqr(x = x, y = y, tau = 0.5, lambda = lambda, lam2 = 0.01,
                  pen = "scad")
nc.preds <- predict(nc.fit, newx = tail(x), s = nc.fit$lambda[3:5])
}
\seealso{
\code{\link{nc.hdqr}}, \code{\link{coef.nc.hdqr}}
}
