% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.hdsvm-methods.R
\name{predict.nc.hdsvm}
\alias{predict.nc.hdsvm}
\title{Make Predictions from a \code{nc.hdsvm} Object}
\usage{
\method{predict}{nc.hdsvm}(object, newx, s = NULL, type = c("class", "loss"), ...)
}
\arguments{
\item{object}{Fitted \code{\link{nc.hdsvm}} object.}

\item{newx}{Matrix of new predictor values at which predictions are to be
made. This is a required argument.}

\item{s}{Values of the penalty parameter \code{lambda} at which
predictions are requested. Default is the entire sequence used during
the model fit.}

\item{type}{Type \code{"class"} (the default) returns the predicted class
labels (\eqn{-1}/\eqn{+1}); type \code{"loss"} returns the linear
predictor.}

\item{...}{Not used.}
}
\value{
A matrix of predicted values, one row per row of \code{newx} and
  one column per value of \code{s}.
}
\description{
Produces class labels or the linear predictor for new predictor data
using a fitted \code{\link{nc.hdsvm}} object at specified \code{lambda}
values.
}
\examples{
set.seed(315)
n <- 100
p <- 400
x1 <- matrix(rnorm(n / 2 * p, -0.25, 0.1), n / 2)
x2 <- matrix(rnorm(n / 2 * p, 0.25, 0.1), n / 2)
x <- rbind(x1, x2)
beta <- 0.1 * rnorm(p)
prob <- plogis(c(x \%*\% beta))
y <- 2 * rbinom(n, 1, prob) - 1
lambda <- 10^(seq(1, -4, length.out = 30))
\donttest{
nc.fit <- nc.hdsvm(x = x, y = y, lambda = lambda, lam2 = 0.01, pen = "scad")
nc.preds <- predict(nc.fit, newx = tail(x), s = nc.fit$lambda[3:5])
}
}
\seealso{
\code{\link{nc.hdsvm}}, \code{\link{coef.nc.hdsvm}}
}
